requirements.txt
selenium
webdriver-manager
flask-cors
flask
flask-limiter
python-decouple
pyodbc
flask-jwt-extended